+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added "Reactions" tab to user's own profile page.
 - Display "in group {group}" subheader if post is addressed to a group.
+- Added "Group members" page.
 
 ## [5.7.1] - 2026-07-12